consolidate auth
